requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
dotenv==0.9.9
python-dotenv==1.0.1
sqlparse==0.5.3
